refactor(config):将原先边界过宽或是职责细碎的部分重构
